remoteproc: qcom: Add driver to bring RPM out of reset
In some firmware configurations Linux might be responsible to bring the
RPM processor out of reset (rather than starting it very early from
boot firmware).

Add a simple driver that allows representing the RPM like a standard
remoteproc in the device tree. When the reset is specified (optional),
the reset control is deasserted during probe. Note that it is not
possible (and intended) to stop the RPM remoteproc again. Also, the
firmware is already loaded by earlier boot firmware so the remoteproc
subsystem is not actually used at the moment.

Signed-off-by: Stephan Gerhold <stephan@gerhold.net>